syslinux: rewrite recipe so only target code is x86-specific
Currently the syslinux recipe set COMPATIBLE_HOST to 32/64-bit x86.
Whilst this makes sense for the target binaries as syslinux is
x86-specific, this also affects the native recipe which should be able
to be built on non-x86 build hosts (for example, arm64 build host,
qemux86 target).

syslinux itself has a somewhat fragile build system but with care it is
possible to build the installer for native/nativesdk, and the bootloader
for target.  Despite claims the documentation to the contrary, not all
of the bootloader is shipped precompiled in the tarball which does
complicate matters.

Signed-off-by: Ross Burton <ross.burton@arm.com>