yocto-check-layer: make argument parsing unambiguous
Currently the argument parser uses nargs=+ for both the layers
(positional arguments) and machines/dependencies/addition layers
(optional arguments).  This means it's impossible to determine what is
meant by:

$ yocto-check-layer --machines a b c

Is this machines=(a,b) and layers=(c), or machines=(a) and layers=(b,c)?
Or even machines=(a,b,c) and layers=(), which results in a parse failure
as the layers are mandatory?

The workaround is to re-order the arguments to an unambiguous form:

$ yocto-check-layer b c --machines a

However this means the user needs to know the problem and how to work
around it.  Instead, change the parser to require an explicit --layers
argument:

$ yocto-check-layer --machines a --layers b c

Also improve the other arguments by adding clearer metavars and short
options.

[ YOCTO #14148 ]

Signed-off-by: Ross Burton <ross.burton@arm.com>