toaster: Filtering out erroneously empty bblayer names.
Within toaster.bbclass on toaster_layerinfo_dumpdata the str.strip is only applied to the value that comes from BBLAYERS. There are chances (also the case that I ran into by mistake) when there is an additional \t which will lead to trying to process a layer with name "\t". IMO this can be accepted and I have added an extra strip on layer name before filtering by empty layer name.

Signed-off-by: Bogdan Ilies <ene-ilies@users.noreply.github.com>