bitbake: Use virtual environment for bitbake
Configure bitbake to run inside a virtual environment. This is done with
a bbpython wrapper script used in place of "python" when executing
scripts. This wrapper will create or update the virtual environment as
necessary, then invoke the underlying script using the python from the
virtual environment.

Users can set the environment variable BITBAKE_VENV if they would like
to choose where the virtual environment gets placed, otherwise it will
default to the venv directory in bitbake

TODO: Several script in OE may need to be updated to use bbpython

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>