util-linux: use the system binaries when running ptest
Unless told otherwise, the util-linux test suite assumes that it is
running in a build tree and looks for binaries to run.

In ptest we're on the target, so it needs to run the system binaries
instead.

This reduces ~180 skipped test cases down to tens, and should fix weird
behaviour.

Also pass --parsable to reduce the amount of noise in the log ouput, and
--show-diff so that the reason for the failure is in the log.

[ YOCTO #14244 ]

Signed-off-by: Ross Burton <ross.burton@arm.com>