meta-skeleton: add recipe examples from documentation sources
This adds three examples previously located in the yocto-docs
repository (under documentation/ref-manual/examples).

The examples were not used for building the documentation
so it's better to keep them here where they can automatically be
tested.

Fixed the "hello-single" to add a version number to the .bb file.
After this, all examples build fine with the "master" branch of Poky.

Note that the "mtd-utils" example, and example of a recipe
for an Autotools page, was not copied over. It doesn't build any
more, and anyway, we already have an official recipe for a recent version
of "mtd-utils" in OE-core.

(From OE-Core rev: 449ddb709b37df37606df439d987904672a4de13)

Signed-off-by: Michael Opdenacker <michael.opdenacker@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>