lttng-modules: update to 2.13.0
Updating to 2.13.0, so we can pick up the latest -stable and
support for 5.14+ kernels.

We refresh one patch for new context, and update the LICENSE
checksum due to commit 2df37e95fa4303 [Cleanup: Move headers from
toplevel to include/lttng/]. which impacted the LICENSE file (but
licensing is the same).

MODULES_MODULE_SYMVERS_LOCATION must also be specified in this
release, as the lttng build has moved the module target to a
'src' subdirectory.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>