rust-cross*: ignore how some sstate sigs are computed
The test:
   sstatetests.SStateTests.test_sstate_sametune_samesigs
which checks if the sstate checksums of two identical machines
(using the same tune) are the same, apart from changes within
the machine specific stamps directory, fails on the assertion:
   self.assertCountEqual(files1, files2)
due to the signature of various 32 bit package builds such as:
   x86_64-linux/lib32-rust-cross-i686
   x86_64-linux/rust-cross-i686
   x86-pokymllib32-linux/lib32-libstd-rs
   x86-pokymllib32-linux/lib32-rust
differing. Jumping down the rabbit hole past all the bitbake-diffsig
outputs that differ due to dependent hashes, you come to a  diff of:
   -Variable MACHINEOVERRIDES value is ${@bb.utils.contains('TUNE_FEATURES', 'mx32', 'x86-x32:', '', d)}${@bb.utils.contains('TUNE_FEATURES', 'm32', 'x86:', '', d)}qemuall:${MACHINE}
   +Variable MACHINEOVERRIDES value is ${@bb.utils.contains('TUNE_FEATURES', 'mx32', 'x86-x32:', '', d)}${@bb.utils.contains('TUNE_FEATURES', 'm32', 'x86:', '', d)}qemuall:${MACHINE}:qemux86
in
   stamps/x86_64-linux/rust-cross-i686/1.54.0-r0.do_rust_gen_target.<sig>

Ignore these signatures by setting the vardepvalue for the variables:
  target_is_armv7 and llvm_features_from_tune
as is done for other cross-toolchains in:
   39bfa0dd32 recipes/*-cross recipes: ignore TARGET_ARCH sstate hash

See the bitbake commit log for:
   6c879b44 data/siggen: Add vardepvalue mechanism to allow the variable dependency code to be forced to specific values
and other usages of vardepvalue in oe-core for background.

Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>