bitbake: siggen: Work around unihash issues
Currently we have a bit of a determinism issue with hashequiv since there
can be a unihash assigned by the server between our original query and
this code reporting the outhash. This is fine if the outhashes are the same
but if the outhashes are different (a determinism issue or a different
arch situation). If the outhashes do differ, build can then proceed in a
different path than a later build doing fresh unihash lookups would proceed
in.

There are changes on the server side we could make to try and address
this but those are invasive this close to release. For now, requery the
server and update the unihash. This should avoid the determinism issues
and match the behaviour some future server changes will have in a way that
is a reasonable workaround for this release.

(Bitbake rev: 21ed048c1ea88831e07ced7d8118fe9e39dbc225)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>