license: Allow treating missing license as error
Use mechanism inspired by insane.bbclass to allow individual recipes or
other configuration to determine whether a missing licence should be
treated as a warning (as it is now) or as an error. This is controlled
by whether the error class is in WARN_LICENSE or ERROR_LICENSE.

Use bb.fatal in the error case to ensure that the task really fails. If
only bb.error is used then do_populate_lic isn't re-run on subsequent
builds which could lead to the error being missed.

(From OE-Core rev: 137e478dec4da842145f0cd292e62000fbc092ae)

Signed-off-by: Mike Crowe <mac@mcrowe.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>