librsvg: update 2.40.21 -> 2.52.0 (transition to rust!)
After some poking and hacking I have arrived to the settings that work.
Please refer to the commit for details.

Some of these issues may re-occur in other components; in that case
we need to come up with more generic solutions or upstream fixes.

So far, it's a learning exercise for everybody in getting to know the rust toolchain.

(From OE-Core rev: edf032acf33e4fefda48cebdd0c8d99d6bd5d018)

Signed-off-by: Alexander Kanavin <alex@linutronix.de>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>