python3-tomli: add recipe for 1.2.1
A lil' TOML parser

This is a runtime dependency for setuptools-scm upgrade as well as for
several other Python Packaging Authority (pypa) tools that will be
replacing distutils and setuptools in the future.

Upstream does not have a setup.py, but rather relies on flit_core to
build. This introduces a circular dependency, as flit_core has a runtime
dependency on tomli.

Borrow a setup.py from Gentoo until a better solution is found by
upstream Python Packaging Authority (pypa).

Signed-off-by: Tim Orling <timothy.t.orling@intel.com>