reproducible: Move variable definitions to bitbake.conf
The reproducibility code was originally developed as separate
standalone class but development is no longer experimental and
the code is widely tested and used by default for poky. Reproducible
builds are the direction we need to take as a project.

Transition the core variable definitions to bitbake.conf as part of
a move to make these part of the default workflow.

This also helps reduce test matrix complexity as there is now one
code path.

(From OE-Core rev: e87d01a63abbf7cd96a477c9bae6d68e1fd6cec5)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>