reproducible: Merge code into base.bbclass
Reproducibility is here to stay and needs to be part of our default workflow.
Move the remaining code to base.bbclass so it is always a first class citizen
and it is clear people need to be mindful of it.

(From OE-Core rev: 0c0d632e61957638516e0cbd98d075870ef99f80)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>