bitbake: fetch2/checksum/siggen: Fix taskhashes not tracking file directories
Currently if you have something like:

SRC_URI = "file://foobar;subdir=${S}"

and a file like:

foobar/1/somefile

and then move it to:

foobar/2/somefile

the task checksums don't reflect/notice this. The file-checksum fields
encode two pieces of data, the file path and whether or not the file
exists. Changing the code which uses these fields is problematic.

We can however add a "/./" path element which means "include the bit
after the marker in the checksum" which the path walking code can use
to mark which bits of the path are visible to the fetcher.

I'm not convinced this is great design but it does appear to work.

(Bitbake rev: 13f1126626300de88376c6a83fb92bf3e4140dd6)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>