poky.yaml: fix lz4 package name for older Ubuntu versions
It turns out that for Ubunut, lz4 is the valid package name for newer
versions (e.g. 20.04 LTS) but not for older ones (e.g. 18.04 LTS, where
the correct package is liblz4-tool). In 20.04 the lz4 package includes
liblz4-tool in its "provides" so it's best to list that one for now.

Signed-off-by: Paul Eggleton <paul.eggleton@microsoft.com>