meson: drop redundant is_debianlike() patch
This patch disables is_debianlike() so that it always returns False in the
name of deterministic builds.  However, the caller default_libdir() then
looks at the host filesystem to decide if /usr/lib64 exists and will
return either lib or lib64, so it still isn't deterministic.

This is only used for the default libdir value and meson.bbclass
always passes an explicit --libdir, so this is moot.

Signed-off-by: Ross Burton <ross.burton@arm.com>