bitbake: utils: Add disable_network function
Add a function which uses the unshare glibc call to disable networking
in the current process. This doesn't work on older distros/kernels
but will on more recent ones so for now we simply ignore the cases we
can't execute on. uid/gid can be passed in externally so this can
work with pseudo/fakeroot contexts.

(Bitbake rev: c3d1f04ec49898e09203cbd0dbbcdd9a335ccb11)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>