bitbake: bitbake-worker: Respect nonetwork task flag
Add a "nonetwork" task specific flag which then triggers networking to
be disabled for this task.

This needs to happen before we enter the fakeroot environment of the task
due to the need for the real uid/gid which we save in the parent process.

(Bitbake rev: de0c9f825c3d21000f1f81b6732c037f1c723ea9)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>