python3-flit-core: add recipe for 3.6.0
This is the core of one of the "new build system" tools, although not
officially part of the pypa (Python Packaging Authority) repositories,
it is an increasingly common build tool (e.g. typing_extensions and
tomli) as declared in pyproject.toml for said packages.

This package provides a very simple bootstrapping method that builds
the source tarball (build_sdist) and a wheel (build_wheel).

Wheels are the official vehicle for delivering Python packages now.
Eggs (egg-info) are deprecated and will increasingly go away.

Signed-off-by: Tim Orling <tim.orling@konsulko.com>