flit_core.bbclass: add helper for newer python packaging
Some python packages now use pyproject.toml and declare
flit_core.buildapi as the build engine

Use pip to install, but tell it to not check the index on pypi.org
(--no-index) and not to check for any dependencies (--no-deps).

While the dependency checking is handy, it is flaky in our environment
and occasionally cannot resolve a dependency even though it is present.

Signed-off-by: Tim Orling <tim.orling@konsulko.com>