lib/oe/packagedata.py: add find_file_rprovides
Similar to change to oe-pkgdata-util script.

Look in pkgdata for FILERPROVIDES of the form "perl(DBD::SQLite)"
as returned by the rpm perl.req script.

This saves the step of trying to guess the filepath/filename. We
can instead definitively determine which package provides the
dependency (as long as the provider has been built and packaged).

Signed-off-by: Tim Orling <tim.orling@konsulko.com>