kernel-module-split: modules should recommend the kernel image
Currently each split-out kernel module RDEPENDS on the top-level kernel
package (e.g. kernel-5.15-yocto-standard). Whilst at first this seems
correct, as modules obviously need their matching kernel, there are many
situations where the kernel is provided out-of-band and forcing the
kernel in via RDEPENDS in the wrong thing to do, for example an
initramfs really shouldn't contain a kernel image, but can contain
kernel modules.

Change the module splitting logic to use RRECOMMENDS instead of
RDEPENDS, and tighten the dependency to kernel-image instead of kernel
to pull in just the image, which also means PACKAGE_EXCLUDE =
"kernel-image-*" is sufficient to ensure the image doesn't get pulled
into an image.

Signed-off-by: Ross Burton <ross.burton@arm.com>