[WIP] setuptools_build_meta.bbcclass: PYPA_WHEEL
The setuptools.build_meta.build_wheel backend emits the file name of the
wheel after compilation. Rather than brittle, hacky ways of determining
the file name of the wheel, set PYPA_WHEEL to the output of the backend.

Use a python3 do_compile() function to build the wheel and set the
PYPA_WHEEL variable.

For some reason, this variable is then None in do_install.

Signed-off-by: Tim Orling <tim.orling@konsulko.com>