binutils-cross/gcc-cross: use TARGET_SYS in PN instead of TARGET_ARCH
Building a cross-canadian toolchain requires building cross recipes
that target the canadian system as well (needed to produce libgcc
for the canadian system, for instance).

This requires properly namespacing the cross recipes, e.g. binutils-cross-x86_64-poky-linux
for the main target, and binutils-cross-arm-none-eabi for the canadian one.
Note that the actual content of the recipes has been TARGET_SYS-specific for
a long time, the binary filenames contain that directly.

Signed-off-by: Alexander Kanavin <alex@linutronix.de>