Require all bad licenses to have exceptions to allow a package be built
Previously it was possible to only whitelist one of the incompatible
licenses for a package to allow it to be installed in an image. But the
correct thing to do is to require that all incompatible licenses for a
package must have exceptions to allow it. Otherwise we could just as
well specify only the list of packages that should be allowed to be
installed regardless of their licenses.

Signed-off-by: Peter Kjellerstedt <peter.kjellerstedt@axis.com>