knotty.py: Unify quiet and non-quiet output for the running tasks
This adds the number of currently running tasks to the output in quiet
mode, and removes the word "currently" from the output for the non-quiet
output. Removing "currently" avoids having the text "jump around"
depending on whether any tasks are running or not, and it also gives a
little bit more space to the progress bar.

Signed-off-by: Peter Kjellerstedt <peter.kjellerstedt@axis.com>