python3-setuptools3: clean up PYPA_WHEEL usage
There's no need to set PYPA_WHEEL as the default is sufficient.

Remove the use of PYPA_WHEEL in the native do_install() as this variable
will be disappearing shortly.

Remove the bbfatal_log in the native do_install(), if this breaks then
something has gone very wrong and the user is not expected to fix it.

Signed-off-by: Ross Burton <ross.burton@arm.com>