python3-wheel: clean up PYPA_WHEEL usage
Remove the use of PYPA_WHEEL in the native do_install() as this variable
will be disappearing shortly.

Remove the bbfatal_log in the native do_install(), if this breaks then
something has gone very wrong and the user is not expected to fix it.

Also flit_core inherits setuptools3-base, so no need to inherit it again.

Signed-off-by: Ross Burton <ross.burton@arm.com>