scripts/runqemu: Fix memory limits for qemux86-64
When setting memory to 4GB, qemu is only running with 2GB for x86_64.
Avoid this by removing the mem= option to the kernel and letting the
qemu configuration handle it for x86 in a similar way to mips.

(From OE-Core rev: 52edbbe00ae1261fec42ba894a9c2a1bca0d8e99)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>